refactor(policies): migrate Action and DecisionContext imports to pydantic_models, remove deprecated schema file
